+++ NWaves/Filters/PreEmphasisFilter.cs
@@ -16,7 +16,7 @@ namespace NWaves.Filters
         /// Constructor computes simple 1st order kernel
         /// </summary>
         /// <param name="a">Pre-emphasis coefficient</param>
-        public PreEmphasisFilter(double a = 0.97) : base(new [] {1, -a})
+        public PreEmphasisFilter(double a = 0.97) : base(new [] { 1, -(float)a })
         {
         }
